backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# OSINT Tools - WHOIS and DNS
python-whois==0.8.0
dnspython==2.4.2

# HTTP Client for API calls
aiohttp==3.9.1
httpx==0.26.0
requests==2.31.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# PDF Report Generation
reportlab==4.0.9
pillow==10.2.0

# HTML Parsing and Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# GitHub API
PyGithub==2.1.1

# Email Validation
email-validator==2.1.0
py3-validate-email==1.0.13

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
